Move the boundary search to the map
The boundary search ('search by area') controls are located on the map;
however, previously, the search button which turns on those controls
was located in the sidebar. This moves the button to the map to be
colocated with the controls.
